1.进度条实现调整,增加ProgressItem进度项(还有首次loading文本显示问题并且原根据Java开发模式进度条刷新暂未充分验证) 2.所有timeout和delay的地方全部改成Long类型,时间单位为毫秒(之前是秒,影响面较大)

+++ dfish/dfish-ui/src/main/java/com/rongji/dfish/ui/AbstractDialog.java
@@ -36,7 +36,7 @@ public abstract class AbstractDialog<T extends AbstractDialog<T>> extends Abstra
     protected String snap;
     protected String snaptype;
     protected Integer indent;
-    protected Integer timeout;
+    protected Long timeout;
     protected Widget<?> node;
     protected Boolean moveable;
     protected Boolean fullscreen;
@@ -115,11 +115,11 @@ public abstract class AbstractDialog<T extends AbstractDialog<T>> extends Abstra
         return (T) this;
     }
 
-    public Integer getTimeout() {
+    public Long getTimeout() {
         return timeout;
     }
 
-    public T setTimeout(Integer timeout) {
+    public T setTimeout(Long timeout) {
         this.timeout = timeout;
         return (T) this;
     }
